docs: add docarray, update code for 2.6+
