+++ basics/0_does_nothing/app.py
@@ -1,4 +1,7 @@
-from jina import Document, DocumentArray, Flow
+from jina import Flow
+from docarray import Document, DocumentArray
+
+print("This example is to test that the most basic Jina stuff works without issues. The example isn't intended to do anything at all beyond that")
 
 docs = DocumentArray(
     [
